+++ packages/kit/src/views/Perp/components/usePerpTokenUrlSync.desktop.ts
@@ -1,0 +1,3 @@
+export function usePerpTokenUrlSync(): void {
+  // No-op for non-web platforms
+}
